+++ src/main/kotlin/com/dh/baro/product/application/ProductFacade.kt
@@ -1,8 +1,8 @@
 package com.dh.baro.product.application
 
-import com.dh.baro.product.domain.CategoryService
+import com.dh.baro.product.domain.service.CategoryService
 import com.dh.baro.product.domain.Product
-import com.dh.baro.product.domain.ProductService
+import com.dh.baro.product.domain.service.ProductService
 import com.dh.baro.product.presentation.dto.ProductCreateRequest
 import org.springframework.stereotype.Service
