+++ src/main/kotlin/com/dh/baro/product/domain/repository/CategoryRepository.kt
@@ -1,5 +1,6 @@
-package com.dh.baro.product.domain
+package com.dh.baro.product.domain.repository
 
+import com.dh.baro.product.domain.Category
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CategoryRepository : JpaRepository<Category, Long>
