+++ src/test/kotlin/com/dh/baro/product/domain/ProductQueryServiceTest.kt
@@ -2,6 +2,9 @@ package com.dh.baro.product.domain
 
 import com.dh.baro.core.ErrorMessage
 import com.dh.baro.core.SliceResponse
+import com.dh.baro.product.domain.repository.CategoryRepository
+import com.dh.baro.product.domain.repository.ProductRepository
+import com.dh.baro.product.domain.service.ProductQueryService
 import com.dh.baro.product.presentation.dto.ProductListItem
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.assertions.throwables.shouldNotThrowAny
